+++ src/components/Home.js
@@ -1,18 +1,6 @@
-// import { useNavigate } from 'react-router-dom';
-// import { useContext } from 'react';
-// import AuthContext from '../context/AuthProvider';
-import Navbar from './Navbar';
+import Navbar from './navigation/Navbar';
 
 const Home = () =>
-// const { setAuth } = useContext(AuthContext);
-// const navigate = useNavigate();
-
-// const logout = async () => {
-//   // if used in more components, this should be in context
-//   // axios to /logout endpoint
-//   setAuth({});
-//   navigate('/linkpage');
-// };
   (
     <section className="h-screen">
       <Navbar />
